# swagger.yaml
...
openapi: 3.0.0
info:
  title: CarHexa
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /driver/:
    post:
      summary: CRUD ACTION create
      tags:
        - Driver
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                Address:
                  type: string
                FisrtName:
                  type: string
                LastName:
                  type: string
                License:
                  type: string
                Ratings:
                  type: string
    get:
      summary: CRUD ACTION list
      tags:
        - Driver
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Driver"
  /driver/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Driver
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Driver
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
    post:
      summary: CRUD ACTION update
      tags:
        - Driver
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Driver"
  /driver/findBy_passenger/{key}:
    get:
      summary: CRUD ACTION findBy_passenger
      tags:
        - Driver
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Driver"
    
  #-----  Passenger  -----
  /passenger/:
    post:
      summary: CRUD ACTION create
      tags:
        - Passenger
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Passenger"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                Address:
                  type: string
                FirstName:
                  type: string
                LastName:
                  type: string
                PhoneNumber:
                  type: string
    get:
      summary: CRUD ACTION list
      tags:
        - Passenger
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Passenger"
  /passenger/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Passenger
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Passenger
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Passenger"
    post:
      summary: CRUD ACTION update
      tags:
        - Passenger
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Passenger"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Passenger"
  /passenger/findBy_drivers/{key}:
    get:
      summary: CRUD ACTION findBy_drivers
      tags:
        - Passenger
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Passenger"
    
  #-----  User  -----
...
  /user/:
    post:
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /user/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    
    
  #-----  User  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Driver:
      properties:
        _id:
          type: string
        Address:
          type: string
        FisrtName:
          type: string
        LastName:
          type: string
        License:
          type: string
        Ratings:
          type: string
      required:
        -  _id
    Passenger:
      properties:
        _id:
          type: string
        Address:
          type: string
        FirstName:
          type: string
        LastName:
          type: string
        PhoneNumber:
          type: string
      required:
        -  _id
    User:
      properties:
        _id:
          type: string
        mail:
          type: string
        name:
          type: string
        password:
          type: string
        roles:
          type: string
        surname:
          type: string
        username:
          type: string
      required:
        -  _id
        - password
        - username